WIP - use attribute for form data,
xml and json are both created on the insert after - their meta keys can be unregistered from rest.
(for now we continue to save the json to the db - swapping to use the meta is a separate case)
at the mo = we have values to XML and in php XML to json - could be lazy and leave it like that or values -> json direct!
the initial values come straight off the attributes - i.e as it should be done
remove the button and the xml field and the rawXML from the form.